+++ src/bin/index.ts
@@ -1,4 +1,25 @@
 #! /usr/bin/env node
 
-/* eslint-disable-next-line no-console */
-console.log("Hello typescript-template-cli!");
+import Yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
+
+import { commands } from './commands';
+
+console.log('TypeScript Template');
+
+async function index() {
+	const yarg = Yargs(hideBin(process.argv));
+
+	for (const cmd of commands) {
+		cmd(yarg);
+	}
+
+	await yarg.parse();
+}
+
+index()
+	.then()
+	.catch((err) => {
+		console.error(err);
+		process.exit(1);
+	});