+++ src/bin/options/index.ts
@@ -1,0 +1,5 @@
+export { name } from './name';
+export { openWith } from './open-with';
+export { org } from './org';
+export { type } from './type';
+export { verbose } from './verbose';