+++ src/bin/options/org.ts
@@ -1,0 +1,9 @@
+export function org() {
+	return [
+		'org',
+		{
+			type: 'string',
+			description: 'organization name'
+		},
+	];
+}