+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [1.0.0] Initial Release
 
 ## Additions
+- [Issue #1] As a user, I want to create a new project